Gradient Descent V1
This code implements from scratch the gradient descent algorithm.
This algorithm aims to be a tool to look for optimal solutions in other algorithms, such as linear regression, logistic regression, quadratic regression and so on.
It is based strongly on some mathematical concepts, such as derivatives, partial derivatives and linear functions.

For day-to-day use, there's no need to implement this algorithm from scratch, since some python packages (such as sklearn) already has it implemented for direct use. The objective of this code is to study how this method works in a more basic level, understanding how sklearn and other such libraries do the gradient descent. Alongside some theoretical and mathematical studies on how this algorithm works, I believe I can understand more about it's limitations, capacities and possible aplications.